pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-support-triage-env"
version = "0.1.0"
description = "OpenEnv: IT support ticket triage (routing, priority, first response)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.2",
  "fastapi>=0.115.0",
  "pydantic>=2.0.0",
  "uvicorn[standard]>=0.24.0",
  "requests>=2.31.0",
  "openai>=1.0.0",
  "websockets>=12.0",
  "pyjwt>=2.12.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
server = "support_triage_env.server.app:main"

[tool.setuptools]
packages = ["support_triage_env", "support_triage_env.server"]

[tool.setuptools.package-data]
support_triage_env = ["**/*.yaml", "**/*.yml"]
